Separate profiles app created. Tests altered to reflect change.
