+++ profiles/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('profiles/', views.profiles_index, name='profiles_index'),
+    path('profiles/<str:username>/', views.profile, name='profile'),
+
+]
